+++ test/common/CallCollectiveForked.hpp
@@ -1,0 +1,11 @@
+#ifndef CALLCOLLECTIVEFORKED_H
+#define CALLCOLLECTIVEFORKED_H
+
+#include <vector>
+
+namespace RcclUnitTesting
+{
+    void callCollectiveForked(int nranks, int collID, const std::vector<int>& sendBuff, std::vector<int>& recvBuff, const std::vector<int>& expected, bool use_managed_mem = false);
+}
+
+#endif
